Add the unirestore tool plus example config file and README.md
Unirestore is a tool that allows for restoring a set of sstables from any architecture to any architecture.
It works by pushing data into all the node/ks/table/upload directories, while monitoring disk space on all
the nodes, and calling refresh when a threshold is reached. It also iterates over files based on their source
host and calls refresh when data from each source host is uploaded, so that there would be no risk of
duplicate filenames incoming from different source hosts.

The data is pushed in using rsync (with optionally settable bandwidth limitation) into each of the cluster
nodes in parallel, to speed things up. For additional optimization, the largest sstables are copied in first,
while the nodes have the most free space.

The script contains code that can deal with token ranges and copy files only into nodes with overlapping
tokens, but it's commented out because scanning the dataset with sstablemetadata was extremely slow. For now
it simply pushes all the files to all the nodes.

Additional documentation is in the README.md file.
Configuration is in the restore_config.yml file, also highly commented.