treewide: allow empty nonwrapping_intervals
Previous implementation of the nonwrapping_interval class
did not allow the notion of an empty interval - it had to be represented
as a disengaged optional: std::optional<nonwrapping_interval>.
Handling empty intervals would be more natural, e.g. an intersection
of two disjoint intervals is simply an empty interval.
Hence, the implementation is now extended with empty intervals,
represented as an interval with end > start. For easier transition,
checking if the interval is empty requires the use of a custom
comparator, since that's how intervals are designed now - they can only
be compared with an externally provided comparator instance.