docker: support an external io properties file
There are times - when a rapid deployment is made, where
the user is interested to cut back on the time it takes
to do iotune since the io capabilities are already known.
Another possible reason to do so is to achieve consistent
results across similar deployments.

This commit adds a cli option to docker:
--io-properties-file <file> that if set, will use the
given io_properties file instead of iotune.